Moved HeroProps type to global components dir
