+++ src/components/landing-hero/components/hero/types.ts
@@ -1,0 +1,13 @@
+/**
+ * Copyright (c) HashiCorp, Inc.
+ * SPDX-License-Identifier: MPL-2.0
+ */
+import { ReactNode } from 'react'
+
+export interface HeroProps {
+	heading: string
+	description: string
+	backgroundSlot?: ReactNode
+	imageSlot?: ReactNode
+	foreground?: 'dark' | 'light'
+}
